Changed IterateSolving Methods to have a Func Parameter to allow for easily reusing the iteration whilst executing a different method instead.
Worked on implementing way to solve Guessing Sudoku by first seeing what the lowest count for potential numbers is. And then letting the computer guess before checking again what the newest lowest count is.

Currently there is no way for the computer to go back to a previous sudoku state in case it guessed wrong and the sudoku isn't filled properly.